whangarei-jobs-agent/requirements.txt
playwright==1.47.*
beautifulsoup4==4.12.*
lxml==5.*
python-dateutil==2.9.*
google-api-python-client==2.149.*
google-auth==2.35.*
google-auth-oauthlib==1.2.*
pyyaml==6.0.*
